Tweak spec to make code match test description
`Entity.first` and `@entity.reload` amount to the same thing, but given the test is about "the existing document" using `reload` seems clearer. Plus something might change in future which makes `Entity.first` not be the entity we are interested in but still pass the spec, resulting in a false positive.