Convert `expect` to `allow` where appropriate
If it's in a `before` block, it must be test setup. If it's test setup, you shouldn't be `expect`ing stuff. Also in some cases we are doing test setup inside the test itself, so change those too (and maybe in the future we can move them into `before` blocks too).