Ensure entities have an identifier which identifies them in the source
We want to guarantee that when we import an entity, it will match the same entity in the database if it has previously been imported. Currently this is not guaranteed because the entity may resolve during one import (and so get a resolved identifier) and not resolve (e.g. due to OC api issue, and so not get a resolved identifier) in another.