Handle case where entity is not upsertable due to multiple identifiers
Imagine we have in the database entity E1 with identifer I1 and E2 with I2. In theory it's possible that E3 comes along with _both_ I1 and I2 (effectively this means E1 and E2 have turned out to be the same entity). In that situation we need to merge E1 and E2 before inserting E3. For now, we will raise an exception.