Always take available entity attributes from the source
Currently only some source attributes are used when creating a non-resolved entity. This was a deliberate decision - source attributes are not as reliable as those which come from OC (via resolving). However they do still have some value, and as long as they are provenanced properly (which will come later) it's clear to the user what's going on.