Fix mailer links
https://openownership.atlassian.net/browse/OO-95

The change introduced in https://github.com/openownership/register/commit/c504d6f282bcc44cd3b6e411d199280b834252db broke the mailer links.

This fix introduces a new `SITE_BASE_URL` env var that is used to set the URL for the mailer. This represents the user facing URL the site is eventually exposed on.

As part of this, we also removed the unused `HOST` env var.