Allow transliteration of certain fields in the web UI
The following pages now support a query param `transliterated` (can be used as `?transliterated` or `?transliterated=true`):

- The entity page
- The entity tree page
- The relationship page
- The search results page

When activated, certain fields are transliterated on the fly.

The various links in the pages above have been hooked up to allow the user to toggle transliteration.

This commit also introduces the concept of "decorators" via the [Draper gem](https://github.com/drapergem/draper). This always us to transparently (to the UI code) handle things like transliteration and provides a better way of exposing models to the UI code.