Handle duplicate entities during PSC import
We will now attempt to utomatically merge duplicate entities during a PSC import. If we encounter a merge of two entities with differing company numbers then we will stop the merge and ignore the import of that line (to be on the safe side and not perform a potentially bad merge).