Clean the test db before the suite starts
This caters for situations where a debugger is used and quit without finishing a spec run (thus not completing the `Database.cleaning` block). In this situation the db could still contain data, which is likely to cause a conflict on the next spec run. This change ensures that before a test run starts the db is cleared out.