UI "hack" to cater for outdated `sample_date` cases
Until we get some more clarity from CH on why this is happening, this UI "hack" will have to do.

We're seeing some cases in the PSC data where the `notified_on` date (which we store as `Relationship#sample_date`) is not updated when a BO relationship has ceased/ended, which means that the `ceased_on` date (which we store as `Relationship#ended_date`) is actually newer. This caused the provenance section to show an older date for the "As of" bit, which looked wrong considering the ended date was newer.

This change just chooses the `ended_date` for the "As of" bit if it's newer than the `sample_date`.