Use the previous ordering of the keys in the OC identifier
The ordering of the keys for any object in MongoDB matters when querying. As a result, it treats the following as two _different_ objects rather than identical:

```
{"jurisdiction_code"=>"gb", "company_number"=>"09971773"}
```

```
{"company_number"=>"09971773", "jurisdiction_code"=>"gb"}
```

This commit goes back to the older ordering of keys (with `jurisdiction_code` first).

Note: we need to separately fix the data that has been created in the meantime with the "wrong" ordering of keys.