Remove the psc:import rake task
This was a secondary rake task to the psc:trigger task, which allowed us to point to a specific file and run the import – something most likely only used for development and testing.

To simplify things, we should just have one way of triggering a full and proper PSC import. If we need to run "special" imports pointing to specific files, etc. then we can do these via a console etc.