PSC Import: add the company number to the self link identifier where possible
This helps us avoid the infamous "self link issue" where the PSC snapshot sometimes contains records where the self link is the same but the underlying company/person data is swapped under the hood (due to data validation issues on their end).

In this fix, if we do have a company number (aka registration number) then add that to the identifier, so that next time if the self link is the same but the company number has changed then we create a brand new entity and update the relationship (instead of overwriting the existing entity with incorrect data).