Rename PscImporter#parse_lines to parse_records
This is more than a simple rename, as it means the importer is no
longer reliant on being given a string line from a snapshot, but
operates directly on a record object parsed from those files. The
JSON parsing and record creation moves into the Chunk import
worker because this is where we already do a bunch of bit shuffling
to de-serialise the data we're trying to import anyway.