Move retrieved_at deserialisation into chunk workers
To be more in keeping with the record deserialisation that happens
in these workers now, I've moved the date/time processing too.

The thinking is, these workers are the boundary between our async
tasks and the necessities of storing data in them in a serialised
fashion, therefore, only they should should care about that
deserialisation.