Lazily loop over chunks of results from DK api
If we don't do this lazily, the memory usage grows and grows, resulting
(at last import) in nearly 2G of memory used. I think this is because,
whilst we enumerate over the results in pages, we're still building up
a full array of results in all_records. Profiling the loop with the
memory_profiler gem showed retained objects from the json gem (and only
that gem). These objects disappear when we add .lazy.