Refactor SK import to use chunked workers
As with the recent changes to the DK import, this updates the Slovak
importer to use the same chunked worker structure. The new :trigger rake
task calls a trigger, which downloads data from the API and queues up
chunk workers to process a set of records (100 at a time by default).
The workers call the ImportTask class, which stores the metadata about
the source and then delegates to the importer to import that set of
records.