Find the child entity which is currently valid in SK records
Previously we assumed that we could always take the first entity, but
examining the data showed that sometimes this entity is no longer valid,
so instead we need to look through them all and find the one that isn't.
At the time of the analysis, there is always a single valid entity, so
we don't need to deal with the case where there are multiple.