Render additional info from OC after page load via AJAX request
Rather than making a request to the OC api during the controller action
for showing an entity, the code now loads the data asynchronosly via
javascript after the page is loaded. This has a couple of benefits:
- The initial page load is quicker, since the user doesn't need to wait
  for the OC api to respond (or worse, timeout) before receiving the
  rest of the page.
- Only clients with javascript will request the additional data, which
  should significantly reduce the amount of requests we're making to the
  OC api, since most crawlers won't run javascript in the first place.

The trade-off is that users without javascript won't see the info at
all. I've decided that's ok because the information is supplementary,
rather than a critical part of the page.

This is implemented as an ajax request to load a rendered html partial
so that:
- I could change as little code as possible, continuing to use
  the existing haml and view helpers to render the info
- We didn't have to expose our OC api key or make a cross-site request
  (if the js had talked to OC directly)
- I didn't have to worry translations or html rendering in the js, which
  would have needed extra libraries and added weight to every page