Upgrade Devise to 4.5.0
For Rails > 5.0 compatibility.

As-per Devise's instructions, I checked the translations file and the
initializers for updates we would need to port over. The initializer had
nothing we needed to change, but I realised that we'd modified the
translation strings file a bit.

In order to make future upgrades easier, I've reset the devise
translation file back to the default (for 4.5.0) and moved our
customisations/additions into the app's file instead.

I've included the updates from Devise 4.4.0 and 4.5.0 which added
autocomplete properties to email and password fields, which seemed to be
the main new feature that we'd like to have.