Update forms to use unified form_with
This is the recommended way to build forms in Rails 5.1.

Changes:

- Replace form_for model with form_with model: model
- Replace form_tag url with form_with url: url
- Update form_tag forms to use the formbuilder methods for fields
- Replace as: params with scope: params in form_for forms
- Add explicit ids to all fields (they're no longer automatic) that have
  labels. (This can be removed in Rails 5.2 as they're automatic again)

Note: I explicitly didn't change
app/views/submissions/entities/_continue_params.html.haml to use the
builder methods because then the fields would be scoped to entity_ which
is not how they're accessed in the controller.

With this change, we no longer need to use the
new_framework_defaults_5_1 migration helper, so it's also removed.