Update form_with usage to use auto field ids
We added ids manually in our upgrade to 5.1 because the automatic id
option wasn't available then. It is in 5.2, so we can strip them out
again.

This is also the last thing we need to fully migrate to the default
settings in 5.2, so we can remove the migration initializer and update
out application.rb

Finally, I noticed that I could add all our merged form_with options
for the submissions entity forms into the helper method, since none were
really special enough to be separate, which tidies up those forms a lot.