Increase OC api timeout for additional info to 2 secs
Now that we only hit the api via async calls from javascript, we can
afford to wait a bit longer for the api to respond. We've decided to
increase it because after adding some tracing we've seen quite a lot of
timeouts (perhaps as much as 50% of calls) but we don't really know if
this is sufficient or what the bounds for normal responses are.