Add basic system tests for entity viewing
Lots of these tests need the same basic context and setup so I spun out
some of the RSpec bits into a shared context.

The tree view spec is hard to do without being tied to the dom, but I've
tried to limit it as much as possible.

Likewise, some of the text I'm testing for result in duplication of view
helpers, but I've tried to make it as minimal (and focussed) as I can.