Display a different message when the OC api times out
Often when the api times out you can get the data by just reloading, but
this isn't in any way obvious to a new user. With the change to surface
api timeouts, we can capture errors in the controller and display a
different message to the user to let them know.

For OO-320