Add created_at and updated_at to persisted and non-embedded models
Similar work to this was done before with Submissions, for OO-23, see:
https://github.com/openownership/register/pull/94 but it was decided
then to only implement created_at and a custom changed_at which
represented a user action that changed a Submission. The rationale was
that we didn't have any valid use case for the timestamps, so why
complicate the code, aka YAGNI.

This was a reasonable decision, but since then we've discovered several
bugs like OO-200, OO-326 and OO-330 where I feel that investigation of
the bug and validation of the fix have been hampered by the lack of this
data. One issue with audit data like this is that you don't know you're
going to need it until you do and then it's usually too late to add it.
In my personal experience I've never *regretted* adding these kinds of
fields, and I think it's even the default now in ActiveRecord and other
ORMs.

Given that, this commit adds the standard Mongoid::Timestamps to every
model we have, where that model is created and persisted directly by the
code. In other words, everything except faux models that don't get
saved and models which are only embedded in others and solely get
updated/created at the same time as their parents.

Note: I've not added the touch: true option to any belongs_to
relationships at this point because I want to keep the timestamps 'pure'
for audit purposes. e.g. I don't want an entity showing as updated when
really we just updated a relationship it's in, so that I can see exactly
when we actually updated the entity itself. This is at odds with how
Rails often recommends you use these timestamps because it relies on
touching to do effective cache key expiration, but we can cross that
bridge when we come to it, as we don't use any caching at the moment.

Note: also that Provenance does not get a separate set of
timestamps because it is embedded in a Relationship at creation time and
never created or updated separately. We can therefore always infer the
provenance's timestamps from the Relationship if we want to.

Note: I had to add these to the excluded fields blacklist for entity
merging, which was a bit of a surprise, so I've raised a ticket to see
if there's a better way we can do that.

These timestamps may indeed be overkill for the submissions models,
where we don't have any obvious need for auditing changes to submitted
data outside of user actions. However, as explained above, I'd prefer to
err on the side of making this data available for future use than
needing to retrofit it once the needs have materialised.

For OO-326

fixup! Add created_at and updated_at to persisted and non-embedded models