Log when EntityResolver overwrites company_number
We have suspicisons that this happens more often than we thought, and
we're overwriting source data with the wrong company from
OpenCorporates. This change allows us to begin to keep an eye on it.

Note that OC perform quite a lot of normalisation on numbers, and so to
make reasonable comparisons we have to do some ourselves. This isn't
perfect (there are still plenty of changes which this misses that we
probably don't care about) but it seems reliable and reduces the number
of log messages raised to a low enough number for us to manually review.

For OO-326

fixup! Log when EntityResolver overwrites name or company_number