Increase MongoDB pool size in dev to match sidekiq workers
Having 50 workers makes running most sidekiq jobs (i.e. imports) in
development impossible, because they overwhelm the default connection
pool of 5 and cause a load of errors. This change raises the mongoid
connection pool size to match production so that we can run imports in
dev as we do on production.