Add user agent blocking to Rack::Attack config
We tried being nice and blocking some bots via robots.txt, but not all
of them respect it, so this ups the ante and blocks them completely by
their user agent.

Following the existing convention for blocking ip addresses, I've used
an env var we can then tweak in production, with the idea that initially
at least this will contain simple strings similar to those you'd put in
the robots.txt. They're matched case-insensitively and anywhere in the
UA string.