Record relationship start_date from UK PSC data
Note - with this we're not totally sure that the sample_date is now
correct, because it can't really be a field for both. We examined
Companies House's forms and documentation to make this judgement and are
fairly sure that 'notified_on' corresponds to the form field which has
help text along the lines of 'When did this person/legal entity become a
PSC', aka a start date of the relationship. We also examined filing
histories and it clearly is not the date a filing was made.