Make updated_at timestamps work with upserts, drop created_at
I discovered that `Mongoid::Timestamps` doesn't play nicely with upsert
because it doesn't trigger the same callbacks that other saving
operations do. This adds a new concern wrapping the `Updated` timestamp
concern in a custom one which calls `set_updated_at` during the
`before_upsert` callback as well.

It's basically impossible to make `created_at` timestamps work at all
with upserts because we can't know ahead of time what will happen (an
update or an insert so the normal Created timestamp concern's approach
of setting it before a save doesn't work. You can't even set it
afterwards because Mongoid doesn't return any information from the
database about what operation an upsert actually resulted in.

I tried various fixes involving `after_upsert` callbacks (won't work
because of the lack of info back from the db, but not very nice even if
it did) and overriding `upsert` to use MongoDB's $setOnInsert
(theoretically works, but is very complex and I had issues with getting
whatever mongodb set back into the ruby instance because mongoid assumes
it already has whatever changes happened in memory). In the end I
decided that it was way too much complication for the eventual gain.
Not least because we're already overriding upsert in Entity, so we'd
need even more overriding to make this work, but also because
`created_at` is the least useful of the two timestamps, and for every
model which has a normal _id, it's embedded in the id anyway.

Rather than leave a confusing created_at field in the db which wasn't
being updated in all cases, I've added a migration to remove it
completely. This means I actually removed the existing created_at from
Submissions::Submission too, but I added a helper method back in to
maintain the old api which just loads the time from the id.

Fixes OO-380