Upgrade Elasticsearch to 6.x
Moves the dev docker container to the latest ES version (6.6) and
updates the client lib to the equivalent 6.x version. This happened to
update Faraday to a different point release, so one spec needed updating
where this version is sent as a UA and mocked with webmock.

Note that we're now using the fully open source version of ES, not the
version with xpack etc, so we don't get annoying license messages in the
dev console, and we don't need to disable those features in the docker
setup.

For OO-182