Output a fixed string when a company name is blank
Because we already use Draper decorators, this seemed like the best
place to put the logic (because it's only a presentational concern).

There's a slight bit of complication because we already define a name
method on the decorator to deal with transliteration, hence why I alias
it before overriding.

Fixes OO-371