Move Slovakian address check into child_entity! method
It makes more sense to have all of the 'validity' checks for child
entities in the same place and logged in a similar way so that we can
keep tabs on them.

This also has the side effect of checking the address on the *current*
child entity, not just the first, which seems much more correct.