Add a check for pagination in SK child entities
We don't know if this will occur, but it seems plausible it could and
we'd want to know about it urgently if it did, hence being cautious.

Because this will occur inside a data import, I don't raise an error
directly (as that would break the whole import and require some manual
intervention to fix), just report an error directly to Rollbar. This is
following the example of the way the api client reports errors in
getting data from the api.