Future-proof elasticsearch mapping types and index names
In ES7 you won't be able to have multiple 'types' of documents in a
single index. We don't currently, but our indexes and mappings are set
up in a way that implies we could. We had a document mapping type of
'entity' within an index called 'open_ownership_register_production'. ES
defaults to using `_doc` for the document mapping type name from v6 and
the elasticsearch-model library follows suit, so we'd either have to set
it back to the old name with an explicit `document_type` call in the
model or rename it.

I chose the latter, creating a new index which is explicit that it
contains entities only, and using the default `_doc` mapping type so
that we're sticking to convention.