Upgrade the dyno size used for the SKImport
When running with a smaller dyno, we had significantly less entities
imported, but no error or other indication. A larger dyno didn't have
this issue. I suspect that if we're too slow retrieving the results,
their api loses the search window or something and effectively ends the
results. In the abscence of time to investigate fully, using a large
dyno by default is an easy solution.