Paginate owned companies in EntitiesController#show
This adjusts the querying/decorating for source relationships so that we
can paginate them with Kaminari and present a paginated list to the
user.

Previously, these relationships were accessed via
entity.relationships_as_source, sorted by a RelationshipsSorter and then
sorted again in the view by name.

This had several problems:
- Using relationships_as_source and then accessing the target entities
  for each relationship results in an N+1 query situation, where each
  entity has to be queried separately to retrieve its details.
- RelationshipsSorter sorted by ended_date, but the view then sorted by
  name, resulting in ended relationships no longer appearing at the
  bottom and a waste of effort in the initial sorting.

Instead, this changes the query for source relationships to use
Relationship.includes instead, so that all related companies are
returned in one query, which is considerable performance improvement.

In addition, RelationshipsSorter now sorts by ended_date *and* name,
which I think was the original intention, which means we don't need to
sort anything in the view.

The sorting is less than ideal, because it means we cannot use
pagination to reduce the number of Relationships we have to get from the
db (MongoDB cannot sort across 'related' collections like a relational
database could). However, since the current numbers (1000s in the worst
case but not 10,000s) don't seem to be *too* slow, this is the
least-worst option to get the UX we want. The time taken to render the
resulting entities is a much bigger chunk of the total processing time.

The pagination is copied directly from the search page, so that the UX
is familiar to the user.