Make RelationshipGraph#select_relationships_recursive more readable
I struggled to understand this recursive function without a detailed set
of notes alongside it, so I made a few changes that I think help with
readability.

Changes:

Change the guard on the recursion base case to check we've seen some
relationships rather than some entities

This code checked that `entities` wasn't empty as a way (I think) of
determining that we'd actually traversed some relationships before
reporting an InferredRelationship somewhere. Since the code protected
by the guard doesn't care about `entities` but does care about
`relationships`, it makes more sense to me to guard that directly.

Rename 'block' to 'stop_condition'

`block` is very generic, just describing what type of thing the
parameter is. In the context of this function though it's clearly a stop
condition for the recursion, so I think it makes it more case to name it
as such.

Rename 'state' to 'inferred_relationships'

Again `state` is very generic and didn't help me understand what was
actually being passed around without following it all around the code to
see where it was assigned or modified. It's an array of
`InferredRelationship` instances though, so `inferred_relationships`
makes that more obvious.

Prefix entities and relationships with seen_

These are again quite generic names, but they mean something specific to
the recursive nature of the function, so with this new name it's more
obvious what their purpose is.

Simplify inferred_relationship creation in the recursion base case

As well as a bit less repeated code, moving what was shared between
both branches of the conditional helps convey what the base case is
doing better (creating a new InferredRelationship). What happens in the
special case when there's only one link in the chain is then lower down
and what's actually different about is more obvious.

Use keyword args

Using keyword args further helps readability when we call
select_relationships_recursive, both from select_relationships and in
the recursion, because it helps make sense of both the initialisation
parameters and the array concatenation for the recursive steps.