Move OC api stubbing into a helper
We're going to use this a lot, so it makes sense to share it, even
though it makes a few specs where we don't want the default slightly
more long-winded.

I decided to remove it from the description of the shared contexts too
since we already do other things (like ES indexing) which aren't
explicitly named there and I could foresee running out of space a bit.