Make person merging non-destructive
Stop using EntityMerger for merging Natural Persons and use
NonDestructivePeopleMerger instead. This is specific to merged people
because we can't trust our people merging process as much as we can
trust merging companies that have unique identifiers.

Note that the merge process is now slightly changed, because we can't
'reduce' a list of entities by destroying them one by one. Instead we
just pick the first one as the 'master' and merge the others into it
one-by-one. I've also removed EntityMergeDecider from the process,
because it was totally arbitrary for merging people anyway, but it also
doesn't make sense when we've picked our person to keep up front.