Override Relationship#source to return master_entity if present
The other side of making merged entities appear like the entity they've
been merged into - return the master_entity instead of the original one
when looking at a relationship involving that entity as the source.

This is only done for the source property because we're only merging
people in this way and they only appear as sources.