De-dupe displayed relationships between merged people
By merging people, we can end up with effectively duplicate
relationships in the database which results in quite confusing views for
the user (e.g. seemingly identical nodes owning a company in the same
way in the tree view).

This changes that so that we de-dupe relationships where both entities
are the same *and* the interests are the same. It purposely ignores
start and end dates because we have cases in SK data where these are
somewhat spurious (relationships get started and ended just to update
details on the owner).