Add a model for data sources and statistics about them
For OO-390, to start being able to store computed stats about particular
sources and present those in 'dashboards'. The stats we'd like to show
will never be able to be calculated on the fly, so we need somewhere to
store them.

We also need to store some content for the page about data sources,
hence the fields for markdown content.

Each data source is likely to have different stats because they're
specific to the laws etc in the country, therefore we have a way to
record which stats we want to show.