Separate and pluralise grouped and merged people 'notes'
These two bits of text need to make it clearer what's going on, and they
can be more explicit when they're separate.

For merged entities, we can have a single entity in the list, so it also
needs to be pluralised. The grouped entities note always refers to 2 or
more, so it doesn't.