Add a page for showing data sources and their latest stats
Adds a show page for DataSource models, with lots of conditionals to
deal with the currently sparse data.

There's some interplay between the DataSourceStatistic::Types module and
translation strings which I'm not too happy with, but since this is all
a bit of an experiment at the moment, I'm going to see if it causes
problems in the future before adding a more complicated data model to
solve it.

Likewise, the decorator and controller interaction around the indices
for footnotes is a bit annoying, but I couldn't think of a nicer way to
do it.