Add a system test for view data sources and stats
This is a slightly confused system test, because it's testing both the
page and the stats calculation, but the two are intrinsically linked in
the user's view of the site so I just picked one (the view) to hold both
tests.

I refactored some of the helpers out of the calculator unit tests to be
able to set up the test data for this, and had to add some css classes
to make it possible to test what was in the stats table, because
Capybara doesn't have much support for plucking rows/columns out of html
tables.