Separate generating and loading test data for Heroku/Dev
The import of test data is very flakey because it relies on
calling the OC api for every imported entity, which often causes it to
timeout. This has a big negative consequence because Heroku marks any
review app that timed out as not being properly built, which means any
subsequent push will rebuild the whole database again. In cases where we
we've set up large amounts of data, this has meant we've lost several
days' work, or been unable to push new fixes to QA in the same
environment

To resolve this, I've altered the rake task to have one which
*generates* the data, storing it in the db and dumping out some json
files, with another that just *loads* those json files. As a bonus, the
data is now stored in the repo, though I'll wait and see whether that is
truly a bonus or we'd rather move it out to an S3 bucket like before.

The thinking here is that it's still beneficial to run all the importers
and generate the test data from the ground up when something changes,
but the rest of the time we just want to load it into the DB quickly and
repeatably.

In order to achieve this, I also decided to move all of the data
generation into the rake task, rather than having it split between
seeds.rb and postdeploy. Nothing we had in seeds.rb was truly seed data
(e.g. static data the app needed to function), so it makes more sense to
run it in afterwards. It also makes it much easier to know what you're
dumping and loading into the db.