Make Submission entities_count an explicit field
As far as I can tell, this was a bug, because Mongoid's counter_cache
fields work subtly differently to ActiveRecord's so this field wasn't
really working. However, it was also an issue for me in dumping the data
to json and loading it back unless this field was actually listed on the
model.