Don't include statements from dissolved companies in PSC stats
We'd assumed that Companies House would end the statements when a
company was dissolved, but this doesn't seem to be the case. This means
we need to check each company individually, rather than just being able
to count Statements, hence moving the code inside the main loop over all
companies.