Add some specs for our ISO8601::Date#mongoize/#demongoize
I thought about adapting this to allow us to supply date *time* strings
as well as just date strings, because that would help with the BODS
importer where many dates can theoretically include times too.

In the end, I decided to do this in the BODS importer instead, so as not
to confuse things (ISO8601::Date can't be created directly from a string
with time in it, so it would be weird if a mongo field backed by it could
).

However, I wrote some specs for the initializer to work out what it was
doing and make sure I didn't break it, so it seems a shame to waste
them.