Add a first pass BODS import of example data
There are lots of compromises in this, I've basically left our data
model untouched and wrangled the BODS data into it.

This follows the pattern of other importers in being asynchronous,
allowing us to parallelise the import of statements. However BODS is
different from our other sources in that entities and relationships
appear separately and thus can potentially be out of order.

I've gone with the simplest possible method of dealing with out-of-order
statements, which is to simply enqueue another import task for the
statement in question. By the nature of Sidekiq, this will get run
more-or-less at the end of the other statements, so it's a pretty good
proxy for retrying after the rest have finished.

Likewise, I've used translation strings for the various interest types.
Since I don't know which types can have shares and which can't, I've
gone belt-and-braces and added -exact and -range variants for all of
them. Note also that we have a clash with some DK types, which whilst
fine for now, points towards the need to namespace them in the future.

See OO-241 for more details on the rationale for this and the
compromises/future work.