Make sure upserting Entities doesn't override merged_entities_count
By setting a default value, when we do as_document on the entity to
upsert we get the default in the attributes we're upserting. This seems
like quite an issue with upserting, but not something I'd like to delve
into fixing now, hence fixing it by just removing the default.

I also realised that our existing specs weren't really testing this case
properly, so I added some more.