Refactor postdeploy tasks into service objects
There seems to be some fundamental race condition/issue with testing the
rake task directly that is really hard to track down. Therefore, I've
moved all of the code into separate service object classes which can be
more easily tested.

I've also rebuilt the dev data, to account for the lack of a default
value for merged_entities_count.