Refactor handling duplicate entities into Entity model
The PSC importer had useful code for upserting an entity and then
destructively merging it with another if duplicate identifiers are found.

I think initially this was intended as a PSC import specific feature, but
it's quite useful for other scenarios, including the rolling updates
from OpenCorporates that OO-275 requires. It also feels like something
every importer should be doing (although we've not actually had an
unhandled 'DuplicateEntitiesDetected' error in any of them yet).

There's nothing actually PSC specific to the code, and it feels like it
could easily be a direct feature of the Entity model, so I've moved it
inside. It's been slightly renamed given the new context, to try to make
what it does clearer to the caller.

I've added some specs to the Entity model which cover this, but also
left the existing ones in, because they test the specific case where an
entity resolution during the import causes a duplicate. I agree with the
general comment in the existing test that they're not really unit tests
as they involve invoking the entity merger, entity indexer etc, but I
don't quite know how to deal with this code smell yet and mocking it all
out to make the tests 'purer' doesn't feel like it's going to help
any, hence leaving them quite 'integration-y'.