Add 'real' total stat to PSC dashboard and calculate %s from that
As well as adding the new statistic type and a value for it for Q1 2019
I've slightly refactored the way stats on particular data sources were
specified so that they *include* the various totals now. It seems easier
to do that than special case them, so the order they're displayed in
just depends on the order they're listed.