Fix rendering of direct circular ownerships in graph view
Cytoscape + Dagre can't figure out how to draw a tight loop back to the
same node itself, it needs to be told how big a loop to make.

See: https://github.com/cytoscape/cytoscape.js/issues/2042

I tried setting control-point-distance too, but I couldn't find any value
that worked. The control-point-step-size needs to be big enough to deal
with the largest possible node, which means it looks a bit rubbish on
smaller nodes, but it's usable.