Add raw data models
The idea is that each record from a source (as the source defines a
'record') is stored as a parsed Hash in the db, which is then linked
to the particular Import(s) it was seen in and, via RawDataProvenance
records, the Entity/Relationship/Statement that it resulted in.

This gives us both a record of the raw data as we saw it through the
snapshots of our imports, and an audit trail of what raw data was used
at what point to make up a finalised record.

RawDataRecords are supposed to be de-duplicated via an etag, which will
have to be generated in a source-specific way (i.e. by the import
process) if the source itself doesn't provide one. Initially we're
trialling this with the PSC data, which provides one we've already
tested and consider reliable.

I've gone with storing data as a parsed Hash field rather than something
more 'raw' (i.e. a string). This is for two main reasons:

1. All of the data we currently get (and foresee getting in the future)
   is JSON (so we won't be forcing other data into an unnatural
   representation).
2. Storing it as a Hash gives us the potential to query it in the DB.

I considered the risk of losing something important during the
deserialisation, but decided it didn't outweigh the benefits of querying
(or the costs of storing both the serialised and unserialised forms).

RawDataProvenance should eventually replace the old 'Provenance' model,
but I've added it separately to avoid having to migrate existing
Provenance data into it immediately and dealing with the changes in views
and data exports that would result. When we do it, we can probably
create some back-dated Imports from the dates in Provenance and code
things such that we can deal with a RawDataProvenance having an empty
list of raw_data_chunks.