Update PscImportTrigger to get config from DataSource
Since we already store all the info we need in the DataSource, it makes
sense to just load it rather than hard code it in this class, and this
is a step towards storing records of the import itself too.

I added a spec to test the regex parsing in this class as well since it
was untested before.