Dump raw data models for the PSC import in dev data
So that we have a set of test data that matches real imports.

Note that this requires a data source, but I haven't bothered to load
all the content into it.

Since I nearly forgot to update the associated test to include these
models, I also refactored things so that it's more automatic to dump
the right models.