Port other existing javascripts to webpack-able modules
Mostly just wrapping them in a class (for no particular reason, it just
seemed concise and tidy) and then instantiating that class in the pack.

This involved a few minor changes to the way these js modules interact
with the HTML, mainly to clear up the usage of globals set from inline
JS snippets.